refactor(log): finish the commands/log to workstation promotion
The workstation migration (phase 5 of #890) had already relocated the
render layer plus the state/orchestration cluster (inkInput,
inkViewModel, inkKeymap, inkWorkflows, inkRuntime) into
workstation/runtime/, but two residents were left behind:

- data.ts (getLogRows/getCommitDetail + GitLogRow/GitCommitDetail
  types) is a git data loader, not a command — moved to
  src/git/logData.ts alongside the other *Data.ts modules.
- interactive.ts / interactiveState.ts (the non-TTY snapshot fallback
  and its state slice) render workstation surfaces — moved to
  workstation/runtime/ next to inkRuntime.ts, which is their only
  caller.

commands/log/ now holds only the actual CLI command: handler.ts,
config.ts, render.ts, commitCompose.ts. Pure move — every import site
(50+ files) updated to the new paths, no behavior change.

Updated the workstation README's layout tree and migration notes,
which had drifted further than this one move accounts for: the
state/orchestration files were already documented as still living in
commands/log/ when they'd actually been in workstation/runtime/ for
some time.

Fixes #1638.
