+++ src/commands/log/handler.ts
@@ -3,7 +3,7 @@ import { Config } from '../../lib/config/types'
 import { loadConfig } from '../../lib/config/utils/loadConfig'
 import { isEmptyRepo } from '../../lib/simple-git/isEmptyRepo'
 import { handleResult } from '../../lib/ui/handleResult'
-import { getCommitDetail, getLogRows } from './data'
+import { getCommitDetail, getLogRows } from '../../git/logData'
 import { startCocoUiFromLogArgv } from '../ui/handler'
 import { applyRepoFlag } from '../utils/applyRepoFlag'
 import { formatCommitDetail, formatLogJson, formatLogTable } from './render'
