+++ src/commands/ui/handler.ts
@@ -5,7 +5,7 @@ import { Config } from '../../lib/config/types'
 import { loadConfig } from '../../lib/config/utils/loadConfig'
 import { getRepo } from '../../lib/simple-git/getRepo'
 import { LogArgv, LogOptions } from '../log/config'
-import { GitLogRow, getLogRows } from '../log/data'
+import { GitLogRow, getLogRows } from '../../git/logData'
 import { getStashCommitHashes } from '../../git/stashData'
 import { startInkInteractiveLog } from '../../workstation/runtime/inkRuntime'
 import { applyRepoFlag } from '../utils/applyRepoFlag'
