+++ src/workstation/chrome/historyRows.test.ts
@@ -1,4 +1,4 @@
-import { GitLogRow } from '../../commands/log/data'
+import { GitLogRow } from '../../git/logData'
 import {
     commitGlyphFor,
     formatInkHistoryGraphRow,
