+++ src/workstation/chrome/overviewCache.ts
@@ -3,7 +3,7 @@ import * as fs from 'node:fs'
 import * as os from 'node:os'
 import * as path from 'node:path'
 
-import { GitLogRow } from '../../commands/log/data'
+import { GitLogRow } from '../../git/logData'
 import { writeFileAtomic } from '../../lib/utils/atomicFileWrite'
 
 /**
