+++ src/workstation/runtime/hooks/useHistoryRefresh.ts
@@ -20,7 +20,7 @@ import type * as ReactTypes from 'react'
 import type { SimpleGit } from 'simple-git'
 import type { LogArgv } from '../../../commands/log/config'
 import type { LogInkAction, LogInkState } from '../inkViewModel'
-import { LOG_INTERACTIVE_DEFAULT_LIMIT, getLogRows } from '../../../commands/log/data'
+import { LOG_INTERACTIVE_DEFAULT_LIMIT, getLogRows } from '../../../git/logData'
 import { getStashCommitHashes } from '../../../git/stashData'
 import { buildHistoryRefetchArgv } from '../historyRefetchResolver'
 import { isStaleBootLoadResolve } from '../loadMoreResolver'
