+++ src/workstation/runtime/loadMoreResolver.ts
@@ -1,4 +1,4 @@
-import { LOG_INTERACTIVE_DEFAULT_LIMIT } from '../../commands/log/data'
+import { LOG_INTERACTIVE_DEFAULT_LIMIT } from '../../git/logData'
 
 /**
  * Pure decision helpers for the load-more pagination cluster.
