+++ src/workstation/runtime/types.ts
@@ -35,7 +35,7 @@ import type { WorktreeOverview } from '../../git/statusData'
 import type { WorktreeOverview as WorktreeListOverview } from '../../git/worktreeData'
 import type { ClipboardRunner } from '../../git/historyActions'
 import type { LogArgv } from '../../commands/log/config'
-import type { GitLogRow } from '../../commands/log/data'
+import type { GitLogRow } from '../../git/logData'
 import type { LogInkState, LogInkView } from '../../workstation/runtime/inkViewModel'
 import type { LogInkInputKey } from '../../workstation/runtime/inkInput'
 import type { LogInkContextStatus } from '../chrome/context'
